feat: move vulnerability normalization and license extraction functions to helpers.go
